requirements.txt
# ========== 核心依赖（必须安装）==========

# YOLOv8 姿态检测
ultralytics>=8.0.200

# 计算机视觉
opencv-python>=4.8.0

# 数值计算
numpy>=1.24.0,<2.0.0

# Web框架
flask>=2.3.0

# HTTP请求（Kimi API调用）
requests>=2.31.0

# 环境变量管理
python-dotenv>=1.0.0

# OpenAI API调用
openai==2.31.0

# ========== 可选依赖（增强功能）==========

# 图像处理增强
pillow>=10.0.0

# 数据可视化
matplotlib>=3.7.0

# Word报告生成
python-docx>=0.8.11

# JSON数据处理
pandas>=2.0.0


# ========== 开发工具（可选）==========

# 代码格式化
black>=23.0.0

# 单元测试
pytest>=7.4.0
